Add ignore-row toggle and data storage view
Excludes ignored transactions from spend totals/coverage/dashboard
while keeping them visible in the row list, and adds a storage tab
to inspect per-month row counts/sizes and clear data.

+++ src/dates.ts
@@ -1,8 +1,8 @@
-/** "2026-08" -> "August 2026". */
+/** "2026-08" -> "Aug 2026". */
 export function monthKeyLabel(monthKey: string): string {
   const [year, month] = monthKey.split("-").map(Number);
   return new Date(year!, month! - 1, 1).toLocaleDateString("en-IN", {
-    month: "long",
+    month: "short",
     year: "numeric",
   });
 }
